split into modules and improve concurrency
For writing to the socket, we wrap the handle in an MVar to ensure that no
writes are concurrent. All reads are done only in one thread though, so we
don't require the lock before blocking to read. This will allow for blocking
for a message while concurrently writing out messages, possibly received from
the inner communication system for message replication across servers.