implement GithubSignedReqBody combinator
This combinator reimplements the core functionality of ReqBody (so it will
parse the request body according to a content-type header analysis and hand the
result off to the handler) but it also verifies that the signature of the body
matches the given signature in the X-Hub-Signature header.

In order to compute the signature, GithubSignedReqBody requires that the
servant routing Context contain a GithubKey, which simply wraps an IO action
producing the key as a ByteString. I chose IO here because we could conceivably
fetch the key from a database. In our code, we just read the key from a file
before starting the webserver, and then use a pure IO action to yield the key
to the handler.